minijail: add support for network namespaces
Each container can run in its own network namespace. The manifest
of the container has been augmented with a 'netconf' section that
lists the parameters. The global northstar.toml file has also
been augmented with the base IP address range to use.

For information on configuring, see doc/network.txt.

The bulk of the changes are in a new file in minijail, since
the process must enter the namespace during the jailing setup.
This also avoids (as much as possible) polluting the existing
minijail code base.

Basically, a namespace is created based on the subnet index
specified in the manifest. Once created, an open file descriptor
is kept while the namespace is then removed. The process is
then added to the namespace. This has the benefit of completely
releasing all namespace related resources when the process exits.
The existing minijail code base does not offer this functionality,
and I doubt upstream would be interested.